+++ polar-core/src/lib.rs
@@ -37,4 +37,4 @@ mod validations;
 mod visitor;
 mod vm;
 pub mod warning;
-pub mod bnf;
+pub mod normalize;
